Add process close event and wait pid system call (#23)
* Add current_task to process manager
* Add wait pid
* Update make clean target

+++ tests/src/mocks/include/libk/sys_call.mock.h
@@ -8,8 +8,8 @@ extern "C" {
 #include "libk/sys_call.h"
 
 DECLARE_FAKE_VALUE_FUNC(void *, _sys_page_alloc, size_t);
-DECLARE_FAKE_VOID_FUNC(_sys_proc_exit, uint8_t);
-DECLARE_FAKE_VOID_FUNC(_sys_proc_abort, uint8_t, const char *);
+DECLARE_FAKE_VOID_FUNC(_sys_proc_exit, int);
+DECLARE_FAKE_VOID_FUNC(_sys_proc_abort, int, const char *);
 DECLARE_FAKE_VOID_FUNC(_sys_proc_panic, const char *, const char *, unsigned int);
 DECLARE_FAKE_VALUE_FUNC(int, _sys_proc_getpid);
 DECLARE_FAKE_VOID_FUNC(_sys_register_signals, void *);
